Move debouncePromise to den
Rather than dumping everything random into utils we can instead dump everything random into den. At least here there's one level of structure.
