+++ packages/den/async/index.ts
@@ -5,3 +5,4 @@
 export { default as LazilyInitialized } from "./LazilyInitialized";
 export { default as MutexLocked } from "./MutexLocked";
 export { default as signal, type Signal } from "./signal";
+export { default as debouncePromise } from "./debouncePromise";
